eventBroker post in MyAddon Thread and FXML
E4Test displays the DATA post in a text at the bottom.
Had problems getting scenebuilder to run from within eclipse.
I tried restarting eclipse but it didn't work.
The scenebuilder exe is in downloads and failed when I tried 
to set it in preference/javafx because I had it opened.  It
still didn't work after I exited.  After this push I'll try login in and
out.